Finished file reorganization. Improved communication between environment and organism.
